Add command to migrate the deprecated test run configuration
