--- launch.test.schema.json
@@ -1,142 +1,0 @@
-{
-    "title": "JSON schema for launch.test.json files",
-    "type": "object",
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "properties": {
-      "run": {
-        "$id": "/properties/run",
-        "type": "object",
-        "properties": {
-          "default": {
-            "$id": "/properties/run/properties/default",
-            "type": "string",
-            "title": "Default",
-            "description": "The default run config to pick, leave it out if you don't want to run with the default config",
-            "default": "",
-            "examples": [
-              ""
-            ]
-          },
-          "items": {
-            "$id": "/properties/run/properties/items",
-            "type": "array",
-            "title": "Configs",
-            "items": {
-              "$id": "/properties/run/properties/items/items",
-              "type": "object",
-              "properties": {
-                "name": {
-                  "$id": "/properties/run/properties/items/items/properties/name",
-                  "type": "string",
-                  "title": "Name",
-                  "description": "The Name of the config, should be unique"
-                },
-                "projectName": {
-                  "$id": "/properties/run/properties/items/items/properties/projectName",
-                  "type": "string",
-                  "title": "ProjectName"
-                },
-                "workingDirectory": {
-                  "$id": "/properties/run/properties/items/items/properties/workingDirectory",
-                  "type": "string",
-                  "title": "WorkingDirectory"
-                },
-                "args": {
-                  "$id": "/properties/run/properties/items/items/properties/args",
-                  "type": "array",
-                  "title": "Args"
-                },
-                "vmargs": {
-                  "$id": "/properties/run/properties/items/items/properties/vmargs",
-                  "type": "array",
-                  "title": "VmArgs"
-                },
-                "preLaunchTask": {
-                  "$id": "/properties/run/properties/items/items/properties/preLaunchTask",
-                  "type": "string",
-                  "title": "PreLaunchTask",
-                  "default": "",
-                  "examples": [
-                    ""
-                  ]
-                },
-                "env": {
-                  "$id": "/properties/run/properties/items/items/properties/env",
-                  "type": "object",
-                  "title": "env",
-                  "description": "The extra environment variables for the test run."
-                }
-              }
-            }
-          }
-        }
-      },
-      "debug": {
-        "$id": "/properties/debug",
-        "type": "object",
-        "properties": {
-          "default": {
-            "$id": "/properties/debug/properties/default",
-            "type": "string",
-            "title": "Default",
-            "description": "The default debug config to pick, leave it out if you don't want to debug with the default config",
-            "default": "",
-            "examples": [
-              ""
-            ]
-          },
-          "items": {
-            "$id": "/properties/debug/properties/items",
-            "type": "array",
-            "title": "Configs",
-            "items": {
-              "$id": "/properties/debug/properties/items/items",
-              "type": "object",
-              "properties": {
-                "name": {
-                  "$id": "/properties/debug/properties/items/items/properties/name",
-                  "type": "string",
-                  "title": "Name"
-                },
-                "projectName": {
-                  "$id": "/properties/debug/properties/items/items/properties/projectName",
-                  "type": "string",
-                  "title": "ProjectName"
-                },
-                "workingDirectory": {
-                  "$id": "/properties/debug/properties/items/items/properties/workingDirectory",
-                  "type": "string",
-                  "title": "WorkingDirectory"
-                },
-                "args": {
-                  "$id": "/properties/debug/properties/items/items/properties/args",
-                  "type": "array",
-                  "title": "Args"
-                },
-                "vmargs": {
-                  "$id": "/properties/debug/properties/items/items/properties/vmargs",
-                  "type": "array",
-                  "title": "VmArgs"
-                },
-                "preLaunchTask": {
-                  "$id": "/properties/debug/properties/items/items/properties/preLaunchTask",
-                  "type": "string",
-                  "title": "PreLaunchTask",
-                  "default": "",
-                  "examples": [
-                    ""
-                  ]
-                },
-                "env": {
-                  "$id": "/properties/run/properties/items/items/properties/env",
-                  "type": "object",
-                  "title": "env",
-                  "description": "The extra environment variables for the test debug run."
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-  }
