+++ package.nls.zh.json
@@ -9,6 +9,7 @@
     "contributes.commands.java.test.explorer.debug.config.title": "根据配置文件调试",
     "contributes.commands.java.test.cancel.title": "取消测试任务",
     "contributes.commands.java.test.explorer.refresh.title": "刷新",
+    "contributes.commands.java.test.config.migrate.title": "迁移已弃用的 'launch.test.json' 文件",
     "configuration.java.test.report.position.description": "设定测试报告的显示位置",
     "configuration.java.test.log.level.description": "设定日志级别",
     "configuration.java.test.message.hintForDeprecatedConfig.description": "设定插件是否会对使用弃用的配置文件进行提示",
