+++ src/constants/commands.ts
@@ -28,4 +28,5 @@ export namespace JavaTestRunnerCommands {
     export const SHOW_TEST_OUTPUT: string = 'java.test.show.output';
     export const OPEN_TEST_LOG: string = 'java.test.open.log';
     export const JAVA_TEST_CANCEL: string = 'java.test.cancel';
+    export const JAVA_CONFIG_MIGRATE: string = 'java.test.config.migrate';
 }
