+++ src/constants/dialogOptions.ts
@@ -5,3 +5,4 @@ export const LEARN_MORE: string = 'Learn More';
 export const NEVER_SHOW: string = 'Never Show again';
 export const YES: string = 'Yes';
 export const NO: string = 'No';
+export const OPEN_SETTING: string = 'Open Settings';
