+++ src/runConfigs.ts
@@ -5,12 +5,10 @@ import { BUILTIN_CONFIG_NAME } from './constants/configs';
 
 export interface IExecutionConfig {
     name?: string;
-    projectName?: string;
     workingDirectory?: string;
     args?: any[];
     vmargs?: any[];
     env?: { [key: string]: string; };
-    preLaunchTask?: string;
 }
 
 export interface IExecutionConfigGroup {
